requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
yt-dlp>=2026.03.17
faster-whisper==1.1.0
httpx==0.28.1
requests
openai
python-dotenv
google-genai
anthropic
